variable for llm tools, minor file function rename
