+++ lambda_msg_handler.tf
@@ -129,6 +129,7 @@ resource "aws_lambda_function" "msg_handler" {
       az_openai_api_version       = var.az_openai_handler_vars.api_version
       az_openai_deployment_name   = var.az_openai_handler_vars.deployment_name
       az_openai_asst_instructions = var.az_openai_handler_vars.asst_instructions
+      llm_tools_vars              = jsonencode(var.llm_tools_vars)
     }
   }
 }
