+++ lambda_msg_handler/msg_handlers/openai_handler.py
@@ -8,8 +8,8 @@ import boto3
 import botocore
 
 from slack_sdk.errors import SlackApiError
+from msg_handlers.llm_tools import tools
 from msg_handlers.openai_related.utils import get_asst, ask_asst, complete_chat
-from msg_handlers.asst_tools import tools
 from msg_handlers.slack_related.utils import extract_event_details, reply
 
 logger = logging.getLogger()
@@ -134,7 +134,7 @@ def save_slack_event(slack_channel_id, slack_thread_ts, slack_event_ts, role, co
     return
 
 
-def handler(slack_event, slack_client):
+def handler_via_assistant(slack_event, slack_client):
     '''
     Overall slack message processing function
     Simple pass-on to Chatgpt
@@ -169,7 +169,7 @@ def handler(slack_event, slack_client):
     return
 
 
-def chat_completion_handler(slack_event, slack_client):
+def handler_via_chat_completion(slack_event, slack_client):
     '''
     Overall slack message processing function
     Re-collect thread history and call Az OpenAI chat completion
